requirements.txt
annotated-types==0.7.0
anyio==4.4.0
boto3==1.34.126
botocore==1.34.126
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
dnspython==2.6.1
doppler-sdk==1.3.0
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
gunicorn==22.0.0
h11==0.14.0
http-exceptions==0.2.10
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.5
packaging==24.1
pluggy==1.5.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.6
requests==2.32.3
rich==13.7.1
rutificador==0.2.17
s3transfer==0.10.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
wrapt==1.16.0
